Add files via upload
There was an unusual bug in the relationship between the Trick class and the Round class that was causing the agent to perform improperly when multiple rounds were run in sequence. (It worked fine in a single round game.) I eliminated the Trick class and moved the functionality it had into the play_trick method of the Round object and that seemed to fix the issue.